refactor backend document test
